refactor(api): decompose DarkroomService god object into facade + peer services (#690)
Splits the 2,107-line `darkroom.service.ts` god object (17 deps, 7+ domains)
into a thin delegating facade plus focused peer services under
`endpoints/admin/darkroom/services/`, matching the existing
`darkroom-training.service.ts` precedent. Behaviour-preserving: the
DarkroomController and its spec are unchanged, and the facade keeps the exact
public surface the controller depends on.

New collaborators:
- DarkroomCharacterService — persona CRUD + the single shared
  `findPersonaBySlug`/`requirePersonaBySlug` (replaces the 8 duplicated
  persona-lookup blocks; per-call-site NotFound messages preserved)
- DarkroomAssetService — asset listing/review + approved-asset dataset sync
- DarkroomInfraService — owns the EC2/CloudFront clients (removed from the
  main constructor) with a single `runAws` wrapper for the duplicated AWS
  try/catch boilerplate
- DarkroomPublishingService, DarkroomGenerationService, DarkroomIngestService,
  DarkroomMediaService
- DarkroomStatsService — campaigns + pipeline stats, now reading through new
  aggregation methods on Ingredients/TrainingsService instead of reaching
  into `.prisma.*` directly
- DarkroomTrainingOrchestratorService — training reads + `startTraining`
- DarkroomValueReader util — shared pure value-coercion/mapping helpers;
  the two nested-ternary status derivations are now explicit switch functions

Also: deletes the dead `getqueryStats` alias; drops the unused `brandId`
arg from the training implementation (the facade retains a `_brandId`
placeholder to keep the controller call shape stable); moves the local
EC2/ingest interfaces into `interfaces/`.

Tests: new unit specs for DarkroomValueReader (status-derivation matrix) and
DarkroomCharacterService (shared lookup + NotFound messages). Existing
darkroom.controller.spec.ts and darkroom-training.service.spec.ts pass
unchanged. Scoped type-check + biome green; 70 tests pass.
