+++ apps/server/api/src/endpoints/admin/darkroom/darkroom.module.ts
@@ -9,7 +9,16 @@ import { PersonasModule } from '@api/collections/personas/personas.module';
 import { TrainingsModule } from '@api/collections/trainings/trainings.module';
 import { DarkroomController } from '@api/endpoints/admin/darkroom/darkroom.controller';
 import { DarkroomService } from '@api/endpoints/admin/darkroom/darkroom.service';
+import { DarkroomAssetService } from '@api/endpoints/admin/darkroom/services/darkroom-asset.service';
+import { DarkroomCharacterService } from '@api/endpoints/admin/darkroom/services/darkroom-character.service';
+import { DarkroomGenerationService } from '@api/endpoints/admin/darkroom/services/darkroom-generation.service';
+import { DarkroomInfraService } from '@api/endpoints/admin/darkroom/services/darkroom-infra.service';
+import { DarkroomIngestService } from '@api/endpoints/admin/darkroom/services/darkroom-ingest.service';
+import { DarkroomMediaService } from '@api/endpoints/admin/darkroom/services/darkroom-media.service';
+import { DarkroomPublishingService } from '@api/endpoints/admin/darkroom/services/darkroom-publishing.service';
+import { DarkroomStatsService } from '@api/endpoints/admin/darkroom/services/darkroom-stats.service';
 import { DarkroomTrainingService } from '@api/endpoints/admin/darkroom/services/darkroom-training.service';
+import { DarkroomTrainingOrchestratorService } from '@api/endpoints/admin/darkroom/services/darkroom-training-orchestrator.service';
 import { IpWhitelistGuard } from '@api/endpoints/admin/guards/ip-whitelist.guard';
 import { FilesClientModule } from '@api/services/files-microservice/client/files-client.module';
 import { ComfyUIModule } from '@api/services/integrations/comfyui/comfyui.module';
@@ -44,6 +53,15 @@ import { forwardRef, Module } from '@nestjs/common';
   ],
   providers: [
     DarkroomService,
+    DarkroomAssetService,
+    DarkroomCharacterService,
+    DarkroomGenerationService,
+    DarkroomInfraService,
+    DarkroomIngestService,
+    DarkroomMediaService,
+    DarkroomPublishingService,
+    DarkroomStatsService,
+    DarkroomTrainingOrchestratorService,
     DarkroomTrainingService,
     IpWhitelistGuard,
     ModelRegistrationService,
