+++ apps/server/api/src/endpoints/admin/darkroom/interfaces/darkroom-infra.interface.ts
@@ -1,0 +1,7 @@
+export interface EC2InstanceStatus {
+  instanceId: string;
+  name: string;
+  role: string;
+  state: string;
+  instanceType: string;
+}
