+++ apps/server/api/src/endpoints/admin/darkroom/services/darkroom-publishing.service.ts
@@ -1,0 +1,176 @@
+import { CredentialsService } from '@api/collections/credentials/services/credentials.service';
+import { IngredientsService } from '@api/collections/ingredients/services/ingredients.service';
+import { DarkroomValueReader } from '@api/endpoints/admin/darkroom/services/darkroom-value-reader.util';
+import { ObjectIdUtil } from '@api/helpers/utils/objectid/objectid.util';
+import { FacebookService } from '@api/services/integrations/facebook/services/facebook.service';
+import { InstagramService } from '@api/services/integrations/instagram/services/instagram.service';
+import { TwitterService } from '@api/services/integrations/twitter/services/twitter.service';
+import { EncryptionUtil } from '@api/shared/utils/encryption/encryption.util';
+import {
+  CredentialPlatform,
+  DarkroomReviewStatus as DarkroomReviewStatusEnum,
+} from '@genfeedai/enums';
+import { LoggerService } from '@libs/logger/logger.service';
+import { CallerUtil } from '@libs/utils/caller/caller.util';
+import { getErrorMessage } from '@libs/utils/error/get-error-message.util';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+
+/**
+ * Owns publishing approved darkroom assets out to social platforms.
+ */
+@Injectable()
+export class DarkroomPublishingService {
+  private readonly constructorName: string = String(this.constructor.name);
+
+  constructor(
+    private readonly ingredientsService: IngredientsService,
+    private readonly instagramService: InstagramService,
+    private readonly twitterService: TwitterService,
+    private readonly facebookService: FacebookService,
+    private readonly credentialsService: CredentialsService,
+    private readonly loggerService: LoggerService,
+  ) {}
+
+  /**
+   * Publish an approved asset to social platforms.
+   */
+  async publishAsset(
+    ingredientId: string,
+    organizationId: string,
+    brandId: string,
+    platforms: string[],
+    caption?: string,
+  ): Promise<{
+    success: boolean;
+    results: Record<string, { success: boolean; id?: string; error?: string }>;
+  }> {
+    const caller = `${this.constructorName} ${CallerUtil.getCallerName()}`;
+    this.loggerService.log(caller, {
+      brandId,
+      ingredientId,
+      organizationId,
+      platforms,
+    });
+
+    // Verify ingredient exists and is approved
+    const ingredient = await this.ingredientsService.findOne({
+      _id: ingredientId,
+      isDeleted: false,
+      organization: organizationId,
+    });
+
+    if (!ingredient) {
+      throw new NotFoundException(
+        `Asset "${ingredientId}" not found in organization "${organizationId}"`,
+      );
+    }
+
+    if (
+      !DarkroomValueReader.hasReviewStatus(
+        ingredient.reviewStatus,
+        DarkroomReviewStatusEnum.APPROVED,
+      )
+    ) {
+      throw new BadRequestException(
+        `Asset must be approved before publishing (current status: ${ingredient.reviewStatus})`,
+      );
+    }
+
+    if (!ingredient.cdnUrl) {
+      throw new BadRequestException(
+        'Asset must have a CDN URL to be published',
+      );
+    }
+
+    const results: Record<
+      string,
+      { success: boolean; id?: string; error?: string }
+    > = {};
+
+    // Publish to each platform
+    for (const platform of platforms) {
+      try {
+        let platformId: string;
+
+        switch (platform.toLowerCase()) {
+          case 'instagram': {
+            const igResult = await this.instagramService.uploadImage(
+              organizationId,
+              brandId,
+              ingredient.cdnUrl,
+              caption || '',
+            );
+            platformId = igResult.mediaId;
+            results.instagram = { id: platformId, success: true };
+            break;
+          }
+
+          case 'twitter':
+            platformId = await this.twitterService.uploadMedia(
+              organizationId,
+              brandId,
+              ingredient.cdnUrl,
+              caption || '',
+              'image/jpeg',
+            );
+            results.twitter = { id: platformId, success: true };
+            break;
+
+          case 'facebook': {
+            // Facebook requires pageId and pageAccessToken
+            const fbCredential = await this.credentialsService.findOne({
+              brand: ObjectIdUtil.toObjectId(brandId)!,
+              isDeleted: false,
+              organization: ObjectIdUtil.toObjectId(organizationId)!,
+              platform: CredentialPlatform.FACEBOOK,
+            });
+
+            if (!fbCredential?.accessToken || !fbCredential?.externalId) {
+              throw new Error('Facebook credential not found');
+            }
+
+            const decryptedToken = EncryptionUtil.decrypt(
+              fbCredential.accessToken,
+            );
+
+            platformId = await this.facebookService.uploadImage(
+              fbCredential.externalId,
+              decryptedToken,
+              ingredient.cdnUrl,
+              caption || '',
+            );
+            results.facebook = { id: platformId, success: true };
+            break;
+          }
+
+          default:
+            results[platform] = {
+              error: `Unsupported platform: ${platform}`,
+              success: false,
+            };
+        }
+      } catch (error: unknown) {
+        this.loggerService.error(caller, {
+          error: getErrorMessage(error),
+          ingredientId,
+          platform,
+        });
+        results[platform] = {
+          error: getErrorMessage(error),
+          success: false,
+        };
+      }
+    }
+
+    const allSuccessful = Object.values(results).every((r) => r.success);
+
+    return {
+      results,
+      success: allSuccessful,
+    };
+  }
+}
